requirements.txt
camelot-py[cv]==0.8.2
keyname==0.1.0
pandas==1.1.2
pikepdf==1.19.3
pyyaml==5.3.1
